api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.1
requests==2.31.0
python-multipart==0.0.6
pathlib2==2.3.7
uuid==1.30 